moving make_sample_video to tests/utils/video.py
