+++ tests/integration/cli/test_utils.py
@@ -6,9 +6,9 @@ from datumaro.components.media_manager import MediaManager
 from datumaro.util.scope import on_exit_do, scope_add, scoped
 
 from tests.requirements import Requirements, mark_requirement
-from tests.unit.test_video import make_sample_video  # pylint: disable=unused-import
 from tests.utils.test_utils import TestDir
 from tests.utils.test_utils import run_datum as run
+from tests.utils.video import make_sample_video
 
 
 class VideoSplittingTest:
