+++ tests/unit/test_video.py
@@ -1,7 +1,6 @@
 import os.path as osp
 from unittest import TestCase
 
-import cv2
 import numpy as np
 import pytest
 
@@ -14,26 +13,7 @@ from datumaro.util.scope import Scope, on_exit_do, scope_add, scoped
 
 from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, compare_datasets
-
-
-@scoped
-def make_sample_video(path, frames=4, frame_size=(10, 20), fps=25.0):
-    """
-    frame_size is (H, W), only even sides
-    """
-
-    writer = cv2.VideoWriter(
-        path,
-        frameSize=tuple(frame_size[::-1]),
-        fps=float(fps),
-        fourcc=cv2.VideoWriter_fourcc(*"MJPG"),
-    )
-    on_exit_do(writer.release)
-
-    for i in range(frames):
-        # Apparently, only uint8 values are supported, but not floats
-        # Colors are compressed, but grayscale colors suffer no loss
-        writer.write(np.ones((*frame_size, 3), dtype=np.uint8) * i)
+from tests.utils.video import make_sample_video
 
 
 @pytest.fixture(scope="module")
